working on skills/old changes

+++ Necromancy.Server/Model/Character.cs
@@ -9,6 +9,6 @@ namespace Necromancy.Server.Model
         public float Z { get; set; }
         public byte viewOffset { get; set; }
         public int battlePose { get; set; }
-        public byte charaPose { get; set; }
+        public int charaPose { get; set; }
     }
 }
